dm-lwm2m: move IPSO logic to main since it is app specific
The default LWM2M objects can be managed separately from the rest of the
application logic, so remove the IPSO code from lwm2m.c and move it to
main. This allows the user to change the default IPSO elements enabled
and still reuse the lwm2m logic that handles the other objects,
including firmware update.

Signed-off-by: Ricardo Salveti <ricardo.salveti@linaro.org>