dm-lwm2m: add scripts/gen_cred_partition.py script
This script is used to create a binary blob representing a PSK
credential partition.  This partition is used by the dm-lwm2m
sample when DTLS is enabled.

From the ZMP installation directory, run a command like this:

./zephyr-fota-samples/scripts/gen_cred_partition.py \
  --device-id=deadbeef \
  --device-token=000102030405060708090a0b0c0d0e0f \
  --output=cred.bin

The arguments are as follows.

- "--device-id" is a public identifier for the device.
  This is currently limited to eight hexadecimal characters.
- "--device-token" is a secret, device-specific token value (i.e.,
  the token must be different for each device on the network). This is
  a sequence of 32 two-character hexadecimal values, each representing
  a byte. In the above example, the first byte is 0x00, the second is
  0x01, etc.
- "--output" is the output file which will contain the binary.

Afterward you can flash the resulting binary using a command similar
to the following (example shown is for nRF52 partition layout used
with BLE Nano 2):
pyocd-flashtool -se -t nrf52 --address 0x7f000 cred.bin

NOTE: You need to re-flash the credential partition after flashing
the ZMP build as the device flash is normally erased by that process.

Signed-off-by: Michael Scott <michael@opensourcefoundries.com>