lwm2m: move to upstream DFU framework
Delete src/lib/mcuboot.c and src/lib/flash_block.c, as well as their
associated headers. This functionality now has upstream equivalents.
Enable that in the prj.conf and use it in lwm2m.c. Remove the
vestigial flash device pointer from main.c as it's not appropriate to
put it there (and really hasn't been for a while).

Since the Zephyr APIs are not the same as what we've been using within
this application, some code changes are needed.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>