dm-lwm2m: Don't initialize LwM2M client until network is ready
Some portion of the network layer meta data might be used in
the LwM2M client initialization.  In the case of LTE-M driver,
we want to use the IMEI for part of the end point name.

Let's rework the initial flow of the LwM2M setup, so that we
wait for the network interface to be up before continuing.

Signed-off-by: Michael Scott <michael@opensourcefoundries.com>