Consolidate nRF Kconfig settings
The board-specific Kconfig settings for nRF-based boards are almost
entirely the same, and in places where they differ, it's due to
characteristics of the SoC, not the board. Move them to KCONFIG_ROOT
and prj.conf, eliminating duplicated configuration.

This means we can delete boards/BOARD.conf in cases where it is not
truly needed for board-specific configuration, easing porting effort
for this sample to other nRF boards. (We probably want to consider
making this apply to non-nRF chips too as needed later on).

While we're here, delete the unused nrf52.conf and propagate a
boards/README.txt from dm-hawkbit-mqtt.

Signed-off-by: Marti Bolivar <marti@foundries.io>