scripts/leshan: alter "client" functionality to enable partial match
When viewing mgmt.foundries.io the list of devices can be VERY long at
times.  Add to that all of the endpoints are named: zmp:sn:###### it
makes it very hard to run mass update tests on devices connected to
a public server.

To solve this, let's use --client parameter as a way of partially
matching the endpoint names for which you want to update.

This combined with a new CONFIG_FOTA_ENDPOINT_PREFIX setting can enable
endpoint naming such as:
mike:sn:e0473f6d
or
andy:sn:b761c8a9

An example of running the leshan.py script for a partial match would be:
python3 zmp-samples/dm-lwm2m/scripts/leshan.py -m \
  -host https://mgmt.foundries.io/leshan \
  --client mike:sn: \
  --device 96b_nitrogen \
  --threads 8 \
  -u http://192.168.69.121:8003/dm-lwm2m-96b_nitrogen-signed.bin

Signed-off-by: Michael Scott <mike@foundries.io>