dm-lwm2m: replace application_state w/ generic settings storage
Let's replace the application_state partition (which stores the update
counter data) with generic FCB-based settings storage.  This is not
only a clean generic way of saving state for the FOTA application,
but also in the future LwM2M will use this same partition to store
persisted values from the objects.

Signed-off-by: Michael Scott <mike@foundries.io>