scripts/leshan: start updates after all nodes have downloaded firmware
The Leshan update script doesn't work well for OpenThread nodes.  As each
node is processing the update, they end up rebooting.  This might be a node
that is currently bridging from 1 portion of the mesh to another.  By
rebooting that node, it would effectively stop all the nodes behind it from
downloading the update.

To fix this let's split the firmware process into 2 stages:

1. Download: this is the original process minus the portion where upon
completion of the download we execute the "update" method on the nodes.
Instead we pause and wait for the other nodes to complete their downloads.
This portion of the process is limited by the "--threads" option.  If only 1
thread is specified we have a fully serialized download which happens 1 node
at a time (same as it does today).

2. Apply update: this portion of the process now happens after all nodes have
tried #1.  Only successfully downloaded nodes will be triggered for the apply
update stage.  This portion of the process is *NOT* thread limited: this means
all nodes (no matter how many) will be triggered for update at once.

This change doesn't affect the overall time to process OTAs (unless thread
limited).  The total time is the calculated as follows:
1. The download time of the slowest node
2. That node's time to reboot and apply the firmware
3. Time to reconnect to the network so that the script can see update status.

Signed-off-by: Michael Scott <mike@foundries.io>